Prevent crash when MeterReporter is deallocated before upload
If the client calls start() on a MeterReporter instance but does not keep a strong reference to this instance, then it gets deallocated and if any log is available to upload a crash occurs after 10 seconds.
Not sure this is the best way to fix this -- maybe intentionally creating a retain cycle is preferable so that clients don't have to retain their MeterReporter.